Introduce a match-based alternative for parse_alias.
This match-based implementation does not depend on alloc, so it lifts the
alloc requirement for parse_alias. However, it performs worse than the
current HashMap-based implementation, therefore, if alloc is enabled,
parse_alias uses the HashMap-based implementation instead.

This way, also the new (text) parser doesn't depend on alloc either.